moved calculator
Moved calculator out of components folder and into newly created utils folder. Import routes were updated as well
