+++ src/api/index.js
@@ -24,7 +24,7 @@ export const getSetList = async () => {
     const response = await fetch('https://api.scryfall.com/sets');
     const { data } = await response.json();
     let mastersList = ['dmr', '2x2', 'tsr', '2xm', 'uma', 'a25', 'ima', 'mm3', 'ema', 'mm2', 'mma'];
-    let falseSets = ['tscd', 'j21', 'h1r', 'tsb', '4bb', 'sum', 'fbb'];
+    let falseSets = ['tscd', 'j21', 'h1r', 'tsb', '4bb', 'sum', 'fbb', 'jmp', 'j22'];
     let boosterSets = data.map(({ code, name, set_type, released_at }) => {
         if (set_type === 'core' || set_type === 'expansion' || set_type === 'draft_innovation') {
             if (!falseSets.includes(code)) {
